requirements.txt
pytorch-lightning==1.9.5
python=3.7.12
torchvision==0.15.0.dev20230115
fastai==2.7.10
pandas==1.3.5
pypickle=1.1.0
numpy=1.21.6
transformers==4.30.2
efficientnet-pytorch==0.7.1
scikit-learn==1.0.2
matplotlib==3.5.3
seaborn==0.12.2
umap-learn==0.5.3
tqdm==4.64.1


